requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
exceptiongroup==1.1.2
Faker==19.2.0
fonttools==4.41.1
geographiclib==2.0
geopy==2.3.0
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.7.2
mirakuru==2.5.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
polib==1.2.0
port-for==0.7.1
psutil==5.9.5
psycopg==3.1.9
psycopg2-binary==2.9.6
pyparsing==3.0.9
pytest==7.4.0
pytest-postgresql==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.11
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.1
